Fix decimal input and shares precision for split details

+++ backend/models.py
@@ -1,6 +1,7 @@
+
 from datetime import datetime
 
-from sqlalchemy import Boolean, Column, DateTime, Integer, String
+from sqlalchemy import Boolean, Column, DateTime, Integer, String, Numeric
 
 from database import Base
 
@@ -91,7 +92,7 @@ class ExpenseSplit(Base):
     is_guest = Column(Boolean, default=False)
     amount_owed = Column(Integer) # The amount this user owes
     percentage = Column(Integer, nullable=True) # For percentage splits
-    shares = Column(Integer, nullable=True) # For share splits
+    shares = Column(Numeric(10, 2), nullable=True) # For share splits
 
 class ExpenseItem(Base):
     __tablename__ = "expense_items"
