+++ backend/schemas.py
@@ -27,7 +27,7 @@ class ExpenseSplitBase(BaseModel):
     is_guest: bool = False
     amount_owed: int
     percentage: Optional[int] = None
-    shares: Optional[int] = None
+    shares: Optional[float] = None
 
 # Expense guest schemas (for non-group expenses)
 class ExpenseGuestCreate(BaseModel):
@@ -64,7 +64,7 @@ class ItemAssignment(BaseModel):
 class ItemSplitDetail(BaseModel):
     amount: Optional[int] = None  # For EXACT split (in cents)
     percentage: Optional[float] = None  # For PERCENTAGE split (0-100)
-    shares: Optional[int] = None  # For SHARES split
+    shares: Optional[float] = None  # For SHARES split
 
 class ExpenseItemCreate(BaseModel):
     description: str = Field(..., max_length=200)
@@ -139,7 +139,7 @@ class ExpenseSplitDetail(BaseModel):
     is_guest: bool = False
     amount_owed: int
     percentage: Optional[int] = None
-    shares: Optional[int] = None
+    shares: Optional[float] = None
     user_name: str
 
     class Config:
